Pull buildInputs out into supporting-materials, reduce dependencies.
It would be nice to have a shell.nix which doesn't force all supporting
materials to be built, so we pull out the buildInputs in preparation for reuse.

We also replace the PDF's src: it was the whole directory, which would cause
rebuilds if anything at all changes; it's now only the article.tex file. Hence
if the article changes, it'll trigger a rebuild; if any inputs, like graphs,
etc. change then it'll trigger a rebuild; if anything else changes, in a way
which doesn't alter the inputs (e.g. refactoring), then no rebuild will be
triggered.