Draw means and standard deviations on precision/recall plots
We model precision and recall as binomial distributions, i.e. each time we
generate a conjecture it's like drawing from an urn with some ratio of good and
bad conjectures. That ratio is the "quality" of the algorithm which we want to
measure.

The standard deviations appear smaller than would be expected just from looking
at the points, but this may just be a consequence of our modelling assumptions.

One slight complication: to work out the standard deviations, we need to know
the number of ground-truth theorems for ALL of the runs. Even though this
information only depends on the sample (which is drawn before we run the
exploration), HaskellTE only includes this if the exploration was successful. We
want to treat timeouts as if they're just another 0-result run, so we need to
run the conjectures_for_sample script ourselves to see how many things were
wanted.

This works, but isn't ideal since there's a chance for the version of
TEBenchmark we're using to drift from the one used to do the runs. I've put a
FIXME in this repo and opened and issue in HaskellTE.