Fix fill_between edge artefacts in PDF output
Seems like edges can have artefacts, where PDF renderers may insert gaps between
the line and the body (sounds like Morphic3 would help ;) ).

Tried a few ways to work around this, e.g. setting linewidth to None, etc. but
they don't seem to work.

It looks like edges can't be transparent, so our use of translucent #999999 as
the body was forcing the edge and body to be different (the edge was
opaque #999999).

I've replaced the body colour with opaque #CCCCCC. On the white background, this
looks essentially identical to the translucent #999999, except that grid lines
don't show through (side note: this colour looks the same as the grid lines;
that *might* be a problem, but we can leave that up to reviewers to decide).

With an opaque body, the body and edge can be the same colour, and the renderer
is smart enough to spot that the edge line isn't needed at all. Hence the
artefact problem goes away, since there is no edge to have a gap between.