Import textWidth, to avoid rebuilding
textWidth was a derivation, which depended on the contents of the .tex
files. Editing the .tex files would cause textWidth to rebuild, which would
cause everything that used textWidth to rebuild (e.g. our graphs).

Now we import the textWidth derivation, which gives us a string. Changing the
contents of the .tex files will cause that string to be recalculated, but the
things which depend on it (e.g. our graphs) will not get rebuilt if the string's
value hasn't changed (which it probably won't).