Adjust scale of graphs
We want the two survival plots in a column next to the two failure lines. This
requires that each graph be at most 50% of the width.

We can set the width (from which the height is calculated via the golden ratio)
in the Python code. The larger this is, the more spread-out the graph contents
are.

We can use a scalebox in LaTeX to grow/shrink the graphs. This doesn't affect
the graph's layout, but it does make the whole thing smaller.

We want the width-proportion (in Python) * scale factor (in LaTeX) to be at most
0.5. This version uses 0.75 in Python and 0.66 in LaTeX to get 0.495.

There's still a problem with x-axis labels being cut off and y-axis labels being
a bit trimmed.